+++ platform/jvm/capture/src/main/kotlin/io/bitdrift/capture/LoggerImpl.kt
@@ -29,6 +29,7 @@ import io.bitdrift.capture.events.device.DeviceStateListenerLogger
 import io.bitdrift.capture.events.lifecycle.AppExitLogger
 import io.bitdrift.capture.events.lifecycle.AppLifecycleListenerLogger
 import io.bitdrift.capture.events.lifecycle.EventsListenerTarget
+import io.bitdrift.capture.events.lifecycle.ForegroundStateListener
 import io.bitdrift.capture.events.lifecycle.WindowFocusFlushLogger
 import io.bitdrift.capture.events.performance.BatteryMonitor
 import io.bitdrift.capture.events.performance.DiskUsageMonitor
@@ -174,6 +175,10 @@ internal class LoggerImpl(
                         networkAttributes,
                         FieldProvider { clientAttributes.dynamicFields() },
                     ),
+                initialOotbFieldProviders =
+                    listOf(
+                        FieldProvider { clientAttributes.initialOotbFields() },
+                    ),
                 errorHandler = errorHandler,
                 customFieldProviders = fieldProviders,
             )
@@ -255,6 +260,7 @@ internal class LoggerImpl(
         this.loggerId = loggerId
 
         runtime = JniRuntime(this.loggerId)
+        ForegroundStateListener(this, ProcessLifecycleOwner.get()).start()
         if (sessionReplayTarget is SessionReplayTarget) {
             sessionReplayTarget.runtime = runtime
         }
@@ -440,6 +446,13 @@ internal class LoggerImpl(
         value: String,
     ) {
         CaptureJniLibrary.addLogField(this.loggerId, key, value)
+    }
+
+    override fun updateOotbField(
+        key: String,
+        value: String,
+    ) {
+        CaptureJniLibrary.updateOotbLogField(this.loggerId, key, value)
     }
 
     override fun removeField(key: String) {