+++ test/platform/swift/unit_integration/core/network/helper/NetworkTestEnvironment.swift
@@ -101,6 +101,10 @@ final class NetworkTestEnvironment {
             []
         }
 
+        func initialOotbFields() -> [Field] {
+            []
+        }
+
         func customFields() -> [Field] {
             []
         }